portal/requirements.txt
amqp==2.5.2
astroid==2.3.3
autopep8==1.4.4
Babel==2.7.0
billiard==3.6.1.0
bleach==3.1.1
celery==4.3.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
crypto==1.4.1
cryptography==2.8
cssselect==1.1.0
decorator==4.4.1
defusedxml==0.6.0
dicttoxml==1.7.4
Django==3.0.3
django-celery-results==1.1.2
django-celery-beat==1.5.0
django-enumfield==1.5
django-extensions==2.2.5
django-filter==2.2.0
django-guardian==2.1.0
django-mysql==3.2.0
django-session-timeout==0.0.4
djangorestframework==3.10.3
djangotoolbox==1.8.0
docutils==0.15.2
future==0.18.2
geocoder==1.38.1
geoip2==2.9.0
idna==2.8
importlib-metadata==0.23
IP2Location==8.4.1
isort==4.3.21
keyring==19.3.0
kombu==4.6.6
lazy-object-proxy==1.4.3
lxml==4.4.2
maxminddb==1.5.1
mccabe==0.6.1
more-itertools==7.2.0
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==1.4.5
Naked==0.1.31
oauthlib==3.1.0
packaging==19.2
pip-review==1.0
pkginfo==1.5.0.1
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.5
pyquery==1.4.1
python3-openid==3.1.0
pytz==2019.3
pywin32-ctypes==0.2.0
PyYAML==5.1.2
ratelim==0.1.6
readme-renderer==24.0
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rest-social-auth==3.0.0
selenium==3.141.0
shellescape==3.4.1
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tornado==5.1.1
tqdm==4.40.0
twine==3.1.1
typed-ast==1.4.0
typing==3.7.4.1
urllib3==1.25.7
validators==0.14.0
vine==1.3.0
webencodings==0.5.1
wrapt==1.11.2
zipp==0.6.0
aiodns==2.0.0
beautifulsoup4==4.8.0
dnspython==1.16.0
flake8==3.7.8
XMLTODICT==0.12.0
grequests==0.4.0
mypy==0.740
netaddr==0.7.19
plotly==4.2.1
pytest==5.2.0
ipwhois==1.1.0
pycurl==7.43.0.4
fabric==2.5.0
xhtml2pdf==0.2.4
markdown==3.2
gunicorn==20.0.4
#pandas==1.0.1